requirements.txt
python-dotenv
yt-dlp
discord.py
# Optional: PyNaCl for voice support if you plan to use native voice libraries
# PyNaCl
